Add and configure Jest, move common components to 'common' directory, add one Snapshot test for Jest
