+++ .eslintrc
@@ -6,7 +6,8 @@
   "env": {
     "browser": true,
     "mocha": true,
-    "node": true
+    "node": true,
+    "jest": true,
   },
   "globals": {
   },
