--- src/api/__tests__/ApiClient-test.js
@@ -1,5 +1,0 @@
-import { expect } from 'chai';
-import ApiClient from '../ApiClient';
-
-describe('ApiClient tests', () => {
-});
