+++ src/common/__tests__/ZulipButton-test.js
@@ -1,0 +1,11 @@
+import 'react-native';
+import React from 'react';
+import renderer from 'react-test-renderer';
+import ZulipButton from '../ZulipButton';
+
+it('renders correctly', () => {
+  const tree = renderer.create(
+    <ZulipButton />
+  ).toJSON();
+  expect(tree).toMatchSnapshot();
+});
