+++ src/user/ZulipAccountsView.js
@@ -19,10 +19,10 @@ import {
 
 import ZulipPasswordAuthView from './ZulipPasswordAuthView';
 import ZulipDevAuthView from './ZulipDevAuthView';
-import styles from './styles';
-import ZulipLogo from './ZulipLogo';
-import ZulipError from './ZulipError';
-import ZulipButton from './ZulipButton';
+import styles from '../common/styles';
+import ZulipLogo from '../common/ZulipLogo';
+import ZulipError from '../common/ZulipError';
+import ZulipButton from '../common/ZulipButton';
 
 class ZulipAccountsView extends React.Component {
   constructor(props) {
